+++ projects/packages/image-cdn/src/class-image-cdn-core.php
@@ -342,6 +342,7 @@ class Image_CDN_Core {
 			'/\.cdninstagram\.com$/',
 			'/^(commons|upload)\.wikimedia\.org$/',
 			'/\.wikipedia\.org$/',
+			'/^m\.media-amazon\.com$/',
 		);
 
 		$host = wp_parse_url( $image_url, PHP_URL_HOST );