+++ projects/packages/jetpack-mu-wpcom/src/features/wpcom-block-editor-nux/index.js
@@ -1,0 +1,8 @@
+/*** THIS MUST BE THE FIRST THING EVALUATED IN THIS SCRIPT *****/
+import '../../common/public-path';
+import { register } from './src/store';
+
+import './src/disable-core-nux';
+import './src/block-editor-nux';
+
+register();