+++ projects/packages/my-jetpack/_inc/components/connected-product-card/index.tsx
@@ -21,6 +21,7 @@ import type { FC, ReactNode } from 'react';
 
 interface ConnectedProductCardProps {
 	admin: boolean;
+	recommendation: boolean;
 	slug: string;
 	children: ReactNode;
 	isDataLoading?: boolean;
@@ -35,6 +36,7 @@ interface ConnectedProductCardProps {
 
 const ConnectedProductCard: FC< ConnectedProductCardProps > = ( {
 	admin,
+	recommendation,
 	slug,
 	children,
 	isDataLoading,
@@ -86,7 +88,7 @@ const ConnectedProductCard: FC< ConnectedProductCardProps > = ( {
 		const cardDescription = defaultDescription.replace( /\s(?=[^\s]*$)/, '\u00A0' );
 
 		return (
-			<Text variant="body-small" style={ { flexGrow: 1 } }>
+			<Text variant="body-small" style={ { flexGrow: 1, marginBottom: '1rem' } }>
 				{ cardDescription }
 			</Text>
 		);
@@ -125,6 +127,7 @@ const ConnectedProductCard: FC< ConnectedProductCardProps > = ( {
 			Description={ Description ? Description : DefaultDescription }
 			status={ status }
 			admin={ admin }
+			recommendation={ recommendation }
 			isFetching={ isActivating || isInstalling || isProductDataLoading }
 			isDataLoading={ isDataLoading }
 			isInstallingStandalone={ isInstalling }